backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
distro==1.9.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.17.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==0.3.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==2.17.0
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pyparsing==3.3.2
PyPDF2==3.0.1
python-dotenv==1.2.1
python-magic==0.4.27
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
